Update Dreampi script to support GameCube modems
This adds GameCube modem support. To use it, set up dreampi as usual, connect your GameCube to the dreampi modem, and set up your network settings in PSO. 

The settings are as follows:
- Username: gc
- Password: gc
- Phone Number: 2001

You'll also need to set up your in-game DNS settings to point to a private server, like Sylverant: 138.197.20.130.

When connecting using 2001 and the above settings, you will not show up on dreamcast.online.